+++ connect/amazon/ses/client/src/main/kotlin/org/http4k/connect/amazon/ses/model/Message.kt
@@ -1,9 +1,0 @@
-package org.http4k.connect.amazon.ses.model
-
-data class Message(
-    val subject: Subject,
-    val html: Body? = null,
-    val text: Body? = null
-)
-
-