+++ connect/amazon/ses/client/src/main/kotlin/org/http4k/connect/amazon/ses/model/MessageTag.kt
@@ -1,0 +1,10 @@
+package org.http4k.connect.amazon.ses.model
+
+import com.squareup.moshi.Json
+import se.ansman.kotshi.JsonSerializable
+
+@JsonSerializable
+data class MessageTag(
+    @Json(name = "Name") val name: String,
+    @Json(name = "Value") val value: String
+)