+++ connect/amazon/ses/client/src/main/kotlin/org/http4k/connect/amazon/ses/model/Subject.kt
@@ -1,8 +1,0 @@
-package org.http4k.connect.amazon.ses.model
-
-import dev.forkhandles.values.NonBlankStringValueFactory
-import dev.forkhandles.values.StringValue
-
-class Subject private constructor(value: String) : StringValue(value) {
-    companion object : NonBlankStringValueFactory<Subject>(::Subject)
-}