+++ connect/amazon/ses/client/src/test/kotlin/org/http4k/connect/amazon/ses/HttpSESTest.kt
@@ -5,10 +5,12 @@ import com.natpryce.hamkrest.present
 import org.http4k.connect.amazon.CredentialsProvider
 import org.http4k.connect.amazon.FakeAwsEnvironment
 import org.http4k.connect.amazon.core.model.Region
+import org.http4k.connect.amazon.ses.model.Body
+import org.http4k.connect.amazon.ses.model.Content
 import org.http4k.connect.amazon.ses.model.Destination
 import org.http4k.connect.amazon.ses.model.EmailAddress
+import org.http4k.connect.amazon.ses.model.EmailContent
 import org.http4k.connect.amazon.ses.model.Message
-import org.http4k.connect.amazon.ses.model.Subject
 import org.http4k.core.MockHttp
 import org.http4k.core.Uri
 import org.http4k.hamkrest.hasUri
@@ -21,8 +23,8 @@ class HttpSESTest {
     companion object {
         @JvmStatic
         fun requestUriSource() = listOf(
-            Arguments.of(null, Uri.of("https://email.af-south-1.amazonaws.com")),
-            Arguments.of(Uri.of("http://localhost:8080/"), Uri.of("http://localhost:8080"))
+            Arguments.of(null, Uri.of("https://email.af-south-1.amazonaws.com/v2/email/outbound-emails")),
+            Arguments.of(Uri.of("http://localhost:8080/"), Uri.of("http://localhost:8080/v2/email/outbound-emails"))
         )
     }
 
@@ -40,7 +42,18 @@ class HttpSESTest {
 
         // when
         runCatching {
-            ses.sendEmail(EmailAddress.of("john@example.com"), Destination(), Message(Subject.of("Hello")))
+            ses.sendEmail(
+                fromEmailAddress = EmailAddress.of("john@example.com"),
+                destination = Destination(),
+                content = EmailContent(
+                    simple = Message(
+                        subject = Content("Hello"),
+                        body = Body(
+                            text = Content("world")
+                        )
+                    )
+                )
+            )
         }
 
         // then