+++ connect/amazon/ses/fake/src/main/kotlin/org/http4k/connect/amazon/ses/model.kt
@@ -1,16 +1,13 @@
 package org.http4k.connect.amazon.ses
 
+import org.http4k.connect.amazon.ses.model.Destination
 import org.http4k.connect.amazon.ses.model.EmailAddress
 import org.http4k.connect.amazon.ses.model.Message
-import org.http4k.connect.amazon.ses.model.SESMessageId
-import org.http4k.template.ViewModel
-
-data class SendEmailResponse(val messageId: SESMessageId) : ViewModel
+import org.http4k.connect.amazon.ses.model.RawMessage
 
 data class EmailMessage(
     val source: EmailAddress,
-    val to: Set<EmailAddress>,
-    val cc: Set<EmailAddress>,
-    val bcc: Set<EmailAddress>,
-    val message: Message
+    val destination: Destination,
+    val message: Message?,
+    val rawMessage: RawMessage?
 )